Move factory to Calamari.Aws
